corrige casa decimal na divisão

+++ src/br/com/dio/calculadora/Calculadora.java
@@ -16,7 +16,7 @@ public class Calculadora {
 		
 		int soma = soma(a,b);
 		int subtracao = subtracao(a,b);
-		int divisao = divisao(a,b);
+		float divisao = divisao(a,b);
 		int multiplicacao = multiplicacao(a,b);
 		
 		System.out.println("soma " + soma);
@@ -32,7 +32,7 @@ public class Calculadora {
 	public static int subtracao(int a, int b) {
 		return a - b;
 	}
-	public static int divisao(int a, int b) {
+	public static float divisao(float a, float b) {
 		return a / b;
 	}
 	public static int multiplicacao(int a, int b) {
